Добавление функции: Particle System
К сожалению, полностью автономный движок сделать не получилось. Придется
копаться с шейдерами, но я переживу.

+++ engine/ps_update_technique.h
@@ -14,7 +14,7 @@ public:
 
 	void SetParticleLifetime(float Lifetime);
 
-	void SetDeltaTimeMillis(float DeltaTimeMillis);
+	void SetDeltaTimeMillis(unsigned int DeltaTimeMillis);
 
 	void SetTime(int Time);
